+++ src/models/command/commandSource.ts
@@ -1,0 +1,8 @@
+export enum CommandSource {
+  Unknown,
+  Initial,
+  Automatic,
+  Manual,
+  API,
+  Force,
+}