+++ src/models/command/roomRestoreShutterPositionCommand.ts
@@ -1,0 +1,12 @@
+import { CommandSource } from './commandSource';
+import { BaseCommand } from './baseCommand';
+
+export class RoomRestoreShutterPositionCommand extends BaseCommand {
+  public constructor(
+    source: CommandSource | BaseCommand,
+    public readonly recalc: boolean = false,
+    reason: string = '',
+  ) {
+    super('RoomRestoreShutterPositionCommand', source, reason);
+  }
+}