+++ src/models/command/shutterSunriseUpCommand.ts
@@ -1,0 +1,8 @@
+import { BaseCommand } from './baseCommand';
+import { CommandSource } from './commandSource';
+
+export class ShutterSunriseUpCommand extends BaseCommand {
+  public constructor(source: CommandSource | BaseCommand, reason: string = '') {
+    super('ShutterSunriseUpCommand', source, reason);
+  }
+}