+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Start rework of commands especially with focus on command source and the stacked reason
+
 ## 2.22.12 (2024-01-13)
 
 * (theimo1221) Update packages